// app/components/block/types.tsx
import type { TransactionsRow } from "../transactionRow/types";

export interface BlockHeader {
  previousHash: string;
  merkleRoot: string;
  timestamp: number;
  nonce: number;
  difficulty: number;
}

export interface Block {
  blockNumber: number;
  cachedHash: string;
  header: BlockHeader;
  transactions: TransactionsRow[];
}